@@ -1,74 +1,45 @@
 {
-	"_v": "1.5.0",
+	"_v": "2.0.1",
+	"type": "module",
 	"main": "./built/index.js",
 	"scripts": {
 		"start": "node ./built",
-		"build": "tsc",
+		"start-daemon": "nodemon ./built",
+		"build": "tspc",
 		"test": "jest"
 	},
 	"dependencies": {
 		"@types/chalk": "2.2.0",
-		"@types/lokijs": "1.5.4",
-		"@types/node": "16.0.1",
-		"@types/promise-retry": "1.1.3",
-		"@types/random-seed": "0.3.3",
-		"@types/request-promise-native": "1.0.18",
-		"@types/seedrandom": "2.4.28",
-		"@types/twemoji-parser": "13.1.1",
-		"@types/uuid": "8.3.1",
-		"@types/ws": "7.4.6",
-		"autobind-decorator": "2.4.0",
-		"canvas": "2.8.0",
-		"chalk": "4.1.1",
+		"@types/lokijs": "1.5.14",
+		"@types/node": "20.11.5",
+		"@types/promise-retry": "1.1.6",
+		"@types/random-seed": "0.3.5",
+		"@types/seedrandom": "3.0.8",
+		"@types/twemoji-parser": "13.1.4",
+		"@types/uuid": "9.0.7",
+		"@types/ws": "8.5.10",
+		"canvas": "3.0.0",
+		"chalk": "5.3.0",
+		"formdata-node": "6.0.3",
+		"got": "14.0.0",
 		"lokijs": "1.5.12",
 		"memory-streams": "0.1.3",
 		"misskey-reversi": "0.0.5",
-		"module-alias": "2.2.2",
+		"nodemon": "3.0.3",
 		"promise-retry": "2.0.1",
 		"random-seed": "0.3.0",
 		"reconnecting-websocket": "4.4.0",
-		"request": "2.88.2",
-		"request-promise-native": "1.0.9",
 		"seedrandom": "3.0.5",
-		"timeout-as-promise": "1.0.0",
-		"ts-node": "10.0.0",
-		"twemoji-parser": "13.1.0",
-		"typescript": "4.3.5",
-		"uuid": "8.3.2",
-		"ws": "7.5.2"
+		"ts-patch": "3.1.2",
+		"twemoji-parser": "14.0.0",
+		"typescript": "5.3.3",
+		"typescript-transform-paths": "3.4.6",
+		"uuid": "9.0.1",
+		"ws": "8.16.0"
 	},
 	"devDependencies": {
-		"@koa/router": "9.4.0",
-		"@types/jest": "26.0.23",
-		"@types/koa": "2.13.1",
-		"@types/koa__router": "8.0.4",
-		"@types/websocket": "1.0.2",
-		"jest": "26.6.3",
-		"koa": "2.13.1",
-		"koa-json-body": "5.3.0",
-		"ts-jest": "26.5.6",
-		"websocket": "1.0.34"
 	},
-	"_moduleAliases": {
-		"@": "built"
-	},
-	"jest": {
-		"testRegex": "/test/.*",
-		"moduleFileExtensions": [
-			"ts",
-			"js"
-		],
-		"transform": {
-			"^.+\\.ts$": "ts-jest"
-		},
-		"globals": {
-			"ts-jest": {
-				"tsConfig": "test/tsconfig.json"
-			}
-		},
-		"moduleNameMapper": {
-			"^@/(.+)": "<rootDir>/src/$1",
-			"^#/(.+)": "<rootDir>/test/$1"
-		}
-	}
+	"nodemonConfig": {
+    "ignore": ["memory.json"]
+  }
 }
