@@ -1,38 +1,39 @@
 // AiOS bootstrapper
 
-import 'module-alias/register';
+import process from 'node:process';
+import chalk from 'chalk';
+import got from 'got';
+import promiseRetry from 'promise-retry';
 
-import * as chalk from 'chalk';
-import * as request from 'request-promise-native';
-const promiseRetry = require('promise-retry');
+import 藍 from './ai.js';
+import config from './config.js';
+import _log from './utils/log.js';
+import pkg from '../package.json' with { type: 'json' };
 
-import 藍 from './ai';
-import config from './config';
-import _log from './utils/log';
-const pkg = require('../package.json');
-
-import CoreModule from './modules/core';
-import TalkModule from './modules/talk';
-import BirthdayModule from './modules/birthday';
-import ReversiModule from './modules/reversi';
-import PingModule from './modules/ping';
-import EmojiModule from './modules/emoji';
-import EmojiReactModule from './modules/emoji-react';
-import FortuneModule from './modules/fortune';
-import GuessingGameModule from './modules/guessing-game';
-import KazutoriModule from './modules/kazutori';
-import KeywordModule from './modules/keyword';
-import TimerModule from './modules/timer';
-import DiceModule from './modules/dice';
-import ServerModule from './modules/server';
-import FollowModule from './modules/follow';
-import ValentineModule from './modules/valentine';
-import MazeModule from './modules/maze';
-import ChartModule from './modules/chart';
-import SleepReportModule from './modules/sleep-report';
-import NotingModule from './modules/noting';
-import PollModule from './modules/poll';
-import ReminderModule from './modules/reminder';
+import CoreModule from './modules/core/index.js';
+import TalkModule from './modules/talk/index.js';
+import BirthdayModule from './modules/birthday/index.js';
+import ReversiModule from './modules/reversi/index.js';
+import PingModule from './modules/ping/index.js';
+import EmojiModule from './modules/emoji/index.js';
+import EmojiReactModule from './modules/emoji-react/index.js';
+import FortuneModule from './modules/fortune/index.js';
+import GuessingGameModule from './modules/guessing-game/index.js';
+import KazutoriModule from './modules/kazutori/index.js';
+import KeywordModule from './modules/keyword/index.js';
+import TimerModule from './modules/timer/index.js';
+import DiceModule from './modules/dice/index.js';
+import ServerModule from './modules/server/index.js';
+import FollowModule from './modules/follow/index.js';
+import ValentineModule from './modules/valentine/index.js';
+import MazeModule from './modules/maze/index.js';
+import ChartModule from './modules/chart/index.js';
+import SleepReportModule from './modules/sleep-report/index.js';
+import NotingModule from './modules/noting/index.js';
+import PollModule from './modules/poll/index.js';
+import ReminderModule from './modules/reminder/index.js';
+import CheckCustomEmojisModule from './modules/check-custom-emojis/index.js';
+import AiChatModule from './modules/aichat/index.js';
 
 console.log('   __    ____  _____  ___ ');
 console.log('  /__\\  (_  _)(  _  )/ __)');
@@ -45,16 +46,23 @@ function log(msg: string): void {
 
 log(chalk.bold(`Ai v${pkg._v}`));
 
+process.on('uncaughtException', err => {
+	try {
+		console.error(`Uncaught exception: ${err.message}`);
+		console.dir(err, { colors: true, depth: 2 });
+	} catch { }
+});
+
 promiseRetry(retry => {
 	log(`Account fetching... ${chalk.gray(config.host)}`);
 
 	// アカウントをフェッチ
-	return request.post(`${config.apiUrl}/i`, {
+	return got.post(`${config.apiUrl}/i`, {
 		json: {
 			i: config.i
 		},
 		headers: config.headers,
-	}).catch(retry);
+	}).json().catch(retry);
 }, {
 	retries: 3
 }).then(account => {
@@ -66,6 +74,10 @@ promiseRetry(retry => {
 	// 藍起動
 	new 藍(account, [
 		new CoreModule(),
+		new AiChatModule(),
+		new ReminderModule(),
+		new TalkModule(),
+		new CheckCustomEmojisModule(),
 		new EmojiModule(),
 		new EmojiReactModule(),
 		new FortuneModule(),
@@ -74,7 +86,6 @@ promiseRetry(retry => {
 		new ReversiModule(),
 		new TimerModule(),
 		new DiceModule(),
-		new TalkModule(),
 		new PingModule(),
 		new ServerModule(),
 		new FollowModule(),
@@ -86,7 +97,6 @@ promiseRetry(retry => {
 		new SleepReportModule(),
 		new NotingModule(),
 		new PollModule(),
-		new ReminderModule(),
 	]);
 }).catch(e => {
 	log(chalk.red('Failed to fetch the account'));
