@@ -1,41 +1,40 @@
-import autobind from 'autobind-decorator';
-import * as chalk from 'chalk';
-const delay = require('timeout-as-promise');
+import { bindThis } from '@/decorators.js';
+import chalk from 'chalk';
 
-import 藍 from '@/ai';
-import Friend from '@/friend';
-import { User } from '@/misskey/user';
-import includes from '@/utils/includes';
-import or from '@/utils/or';
-import config from '@/config';
+import 藍 from '@/ai.js';
+import Friend from '@/friend.js';
+import type { User } from '@/misskey/user.js';
+import includes from '@/utils/includes.js';
+import or from '@/utils/or.js';
+import config from '@/config.js';
+import { sleep } from '@/utils/sleep.js';
 
 export default class Message {
 	private ai: 藍;
-	private messageOrNote: any;
-	public isDm: boolean;
+	private note: any;
 
 	public get id(): string {
-		return this.messageOrNote.id;
+		return this.note.id;
 	}
 
 	public get user(): User {
-		return this.messageOrNote.user;
+		return this.note.user;
 	}
 
 	public get userId(): string {
-		return this.messageOrNote.userId;
+		return this.note.userId;
 	}
 
 	public get text(): string {
-		return this.messageOrNote.text;
+		return this.note.text;
 	}
 
 	public get quoteId(): string | null {
-		return this.messageOrNote.renoteId;
+		return this.note.renoteId;
 	}
 
 	public get visibility(): string {
-		return this.messageOrNote.visibility;
+		return this.note.visibility;
 	}
 
 	/**
@@ -50,15 +49,14 @@ export default class Message {
 	}
 
 	public get replyId(): string {
-		return this.messageOrNote.replyId;
+		return this.note.replyId;
 	}
 
 	public friend: Friend;
 
-	constructor(ai: 藍, messageOrNote: any, isDm: boolean) {
+	constructor(ai: 藍, note: any) {
 		this.ai = ai;
-		this.messageOrNote = messageOrNote;
-		this.isDm = isDm;
+		this.note = note;
 
 		this.friend = new Friend(ai, { user: this.user });
 
@@ -70,7 +68,7 @@ export default class Message {
 		});
 	}
 
-	@autobind
+	@bindThis
 	public async reply(text: string | null, opts?: {
 		file?: any;
 		cw?: string;
@@ -82,31 +80,31 @@ export default class Message {
 		this.ai.log(`>>> Sending reply to ${chalk.underline(this.id)}`);
 
 		if (!opts?.immediate) {
-			await delay(2000);
+			await sleep(2000);
 		}
 
-		if (this.isDm) {
-			return await this.ai.sendMessage(this.messageOrNote.userId, {
-				text: text,
-				fileId: opts?.file?.id
-			});
+		const postData = {
+			replyId: this.note.id,
+			text: text,
+			fileIds: opts?.file ? [opts?.file.id] : undefined,
+			cw: opts?.cw,
+			renoteId: opts?.renote
+		};
+
+		// DM以外は普通に返信し、DMの場合はDMで返信する
+		if (this.note.visibility != 'specified') {
+			return await this.ai.post(postData);
 		} else {
-			return await this.ai.post({
-				replyId: this.messageOrNote.id,
-				text: text,
-				fileIds: opts?.file ? [opts?.file.id] : undefined,
-				cw: opts?.cw,
-				renoteId: opts?.renote
-			});
+			return await this.ai.sendMessage(this.userId, postData);
 		}
 	}
 
-	@autobind
+	@bindThis
 	public includes(words: string[]): boolean {
 		return includes(this.text, words);
 	}
 
-	@autobind
+	@bindThis
 	public or(words: (string | RegExp)[]): boolean {
 		return or(this.text, words);
 	}
