@@ -1,8 +1,10 @@
-import autobind from 'autobind-decorator';
+import { bindThis } from '@/decorators.js';
 import { EventEmitter } from 'events';
-import * as WebSocket from 'ws';
-const ReconnectingWebsocket = require('reconnecting-websocket');
-import config from './config';
+import WebSocket from 'ws';
+import _ReconnectingWebsocket from 'reconnecting-websocket';
+import config from './config.js';
+
+const ReconnectingWebsocket = _ReconnectingWebsocket as unknown as typeof _ReconnectingWebsocket['default'];
 
 /**
  * Misskey stream connection
@@ -30,7 +32,7 @@ export default class Stream extends EventEmitter {
 		this.stream.addEventListener('message', this.onMessage);
 	}
 
-	@autobind
+	@bindThis
 	public useSharedConnection(channel: string): SharedConnection {
 		let pool = this.sharedConnectionPools.find(p => p.channel === channel);
 
@@ -44,19 +46,19 @@ export default class Stream extends EventEmitter {
 		return connection;
 	}
 
-	@autobind
+	@bindThis
 	public removeSharedConnection(connection: SharedConnection) {
 		this.sharedConnections = this.sharedConnections.filter(c => c !== connection);
 	}
 
-	@autobind
+	@bindThis
 	public connectToChannel(channel: string, params?: any): NonSharedConnection {
 		const connection = new NonSharedConnection(this, channel, params);
 		this.nonSharedConnections.push(connection);
 		return connection;
 	}
 
-	@autobind
+	@bindThis
 	public disconnectToChannel(connection: NonSharedConnection) {
 		this.nonSharedConnections = this.nonSharedConnections.filter(c => c !== connection);
 	}
@@ -64,7 +66,7 @@ export default class Stream extends EventEmitter {
 	/**
 	 * Callback of when open connection
 	 */
-	@autobind
+	@bindThis
 	private onOpen() {
 		const isReconnect = this.state == 'reconnecting';
 
@@ -98,7 +100,7 @@ export default class Stream extends EventEmitter {
 	/**
 	 * Callback of when close connection
 	 */
-	@autobind
+	@bindThis
 	private onClose() {
 		this.state = 'reconnecting';
 		this.emit('_disconnected_');
@@ -111,7 +113,7 @@ export default class Stream extends EventEmitter {
 	/**
 	 * Callback of when received a message from connection
 	 */
-	@autobind
+	@bindThis
 	private onMessage(message) {
 		if (message.data === 'pong') return;
 		const { type, body } = JSON.parse(message.data);
@@ -140,7 +142,7 @@ export default class Stream extends EventEmitter {
 	/**
 	 * Send a message to connection
 	 */
-	@autobind
+	@bindThis
 	public send(typeOrPayload, payload?) {
 		const data = payload === undefined ? typeOrPayload : {
 			type: typeOrPayload,
@@ -159,7 +161,7 @@ export default class Stream extends EventEmitter {
 	/**
 	 * Close this connection
 	 */
-	@autobind
+	@bindThis
 	public close() {
 		this.stream.removeEventListener('open', this.onOpen);
 		this.stream.removeEventListener('message', this.onMessage);
@@ -181,7 +183,7 @@ class Pool {
 		this.id = Math.random().toString();
 	}
 
-	@autobind
+	@bindThis
 	public inc() {
 		if (this.users === 0 && !this.isConnected) {
 			this.connect();
@@ -196,7 +198,7 @@ class Pool {
 		}
 	}
 
-	@autobind
+	@bindThis
 	public dec() {
 		this.users--;
 
@@ -210,7 +212,7 @@ class Pool {
 		}
 	}
 
-	@autobind
+	@bindThis
 	public connect() {
 		this.isConnected = true;
 		this.stream.send('connect', {
@@ -219,7 +221,7 @@ class Pool {
 		});
 	}
 
-	@autobind
+	@bindThis
 	private disconnect() {
 		this.isConnected = false;
 		this.disposeTimerId = null;
@@ -239,7 +241,7 @@ abstract class Connection extends EventEmitter {
 		this.channel = channel;
 	}
 
-	@autobind
+	@bindThis
 	public send(id: string, typeOrPayload, payload?) {
 		const type = payload === undefined ? typeOrPayload.type : typeOrPayload;
 		const body = payload === undefined ? typeOrPayload.body : payload;
@@ -268,12 +270,12 @@ class SharedConnection extends Connection {
 		this.pool.inc();
 	}
 
-	@autobind
+	@bindThis
 	public send(typeOrPayload, payload?) {
 		super.send(this.pool.id, typeOrPayload, payload);
 	}
 
-	@autobind
+	@bindThis
 	public dispose() {
 		this.pool.dec();
 		this.removeAllListeners();
@@ -294,7 +296,7 @@ class NonSharedConnection extends Connection {
 		this.connect();
 	}
 
-	@autobind
+	@bindThis
 	public connect() {
 		this.stream.send('connect', {
 			channel: this.channel,
@@ -303,12 +305,12 @@ class NonSharedConnection extends Connection {
 		});
 	}
 
-	@autobind
+	@bindThis
 	public send(typeOrPayload, payload?) {
 		super.send(this.id, typeOrPayload, payload);
 	}
 
-	@autobind
+	@bindThis
 	public dispose() {
 		this.removeAllListeners();
 		this.stream.send('disconnect', { id: this.id });
