@@ -1,10 +1,10 @@
-import autobind from 'autobind-decorator';
-import * as loki from 'lokijs';
-import Module from '@/module';
-import Message from '@/message';
-import serifs from '@/serifs';
-import { User } from '@/misskey/user';
-import { acct } from '@/utils/acct';
+import { bindThis } from '@/decorators.js';
+import loki from 'lokijs';
+import Module from '@/module.js';
+import Message from '@/message.js';
+import serifs from '@/serifs.js';
+import type { User } from '@/misskey/user.js';
+import { acct } from '@/utils/acct.js';
 
 type Game = {
 	votes: {
@@ -27,7 +27,7 @@ export default class extends Module {
 
 	private games: loki.Collection<Game>;
 
-	@autobind
+	@bindThis
 	public install() {
 		this.games = this.ai.getCollection('kazutori');
 
@@ -40,7 +40,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async mentionHook(msg: Message) {
 		if (!msg.includes(['数取り'])) return false;
 
@@ -75,14 +75,14 @@ export default class extends Module {
 			postId: post.id
 		});
 
-		this.subscribeReply(null, false, post.id);
+		this.subscribeReply(null, post.id);
 
 		this.log('New kazutori game started');
 
 		return true;
 	}
 
-	@autobind
+	@bindThis
 	private async contextHook(key: any, msg: Message) {
 		if (msg.text == null) return {
 			reaction: 'hmm'
@@ -139,7 +139,7 @@ export default class extends Module {
 	/**
 	 * 終了すべきゲームがないかチェック
 	 */
-	@autobind
+	@bindThis
 	private crawleGameEnd() {
 		const game = this.games.findOne({
 			isEnded: false
@@ -156,7 +156,7 @@ export default class extends Module {
 	/**
 	 * ゲームを終わらせる
 	 */
-	@autobind
+	@bindThis
 	private finish(game: Game) {
 		game.isEnded = true;
 		this.games.update(game);
