@@ -1,10 +1,10 @@
-import autobind from 'autobind-decorator';
-import * as loki from 'lokijs';
-import Module from '@/module';
-import Message from '@/message';
-import serifs, { getSerif } from '@/serifs';
-import { acct } from '@/utils/acct';
-import config from '@/config';
+import { bindThis } from '@/decorators.js';
+import loki from 'lokijs';
+import Module from '@/module.js';
+import Message from '@/message.js';
+import serifs, { getSerif } from '@/serifs.js';
+import { acct } from '@/utils/acct.js';
+import config from '@/config.js';
 
 const NOTIFY_INTERVAL = 1000 * 60 * 60 * 12;
 
@@ -14,14 +14,13 @@ export default class extends Module {
 	private reminds: loki.Collection<{
 		userId: string;
 		id: string;
-		isDm: boolean;
 		thing: string | null;
 		quoteId: string | null;
 		times: number; // 催促した回数(使うのか？)
 		createdAt: number;
 	}>;
 
-	@autobind
+	@bindThis
 	public install() {
 		this.reminds = this.ai.getCollection('reminds', {
 			indices: ['userId', 'id']
@@ -34,7 +33,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async mentionHook(msg: Message) {
 		let text = msg.extractedText.toLowerCase();
 		if (!text.startsWith('remind') && !text.startsWith('todo')) return false;
@@ -64,7 +63,6 @@ export default class extends Module {
 		const remind = this.reminds.insertOne({
 			id: msg.id,
 			userId: msg.userId,
-			isDm: msg.isDm,
 			thing: thing === '' ? null : thing,
 			quoteId: msg.quoteId,
 			times: 0,
@@ -72,13 +70,13 @@ export default class extends Module {
 		});
 
 		// メンションをsubscribe
-		this.subscribeReply(remind!.id, msg.isDm, msg.isDm ? msg.userId : msg.id, {
+		this.subscribeReply(remind!.id, msg.id, {
 			id: remind!.id
 		});
 
 		if (msg.quoteId) {
 			// 引用元をsubscribe
-			this.subscribeReply(remind!.id, false, msg.quoteId, {
+			this.subscribeReply(remind!.id, msg.quoteId, {
 				id: remind!.id
 			});
 		}
@@ -94,7 +92,7 @@ export default class extends Module {
 		};
 	}
 
-	@autobind
+	@bindThis
 	private async contextHook(key: any, msg: Message, data: any) {
 		if (msg.text == null) return;
 
@@ -121,12 +119,11 @@ export default class extends Module {
 			msg.reply(serifs.reminder.doneFromInvalidUser);
 			return;
 		} else {
-			if (msg.isDm) this.unsubscribeReply(key);
 			return false;
 		}
 	}
 
-	@autobind
+	@bindThis
 	private async timeoutCallback(data) {
 		const remind = this.reminds.findOne({
 			id: data.id
@@ -140,28 +137,22 @@ export default class extends Module {
 		if (friend == null) return; // 処理の流れ上、実際にnullになることは無さそうだけど一応
 
 		let reply;
-		if (remind.isDm) {
-			this.ai.sendMessage(friend.userId, {
-				text: serifs.reminder.notifyWithThing(remind.thing, friend.name)
+		try {
+			reply = await this.ai.post({
+				renoteId: remind.thing == null && remind.quoteId ? remind.quoteId : remind.id,
+				text: acct(friend.doc.user) + ' ' + serifs.reminder.notify(friend.name)
 			});
-		} else {
-			try {
-				reply = await this.ai.post({
-					renoteId: remind.thing == null && remind.quoteId ? remind.quoteId : remind.id,
-					text: acct(friend.doc.user) + ' ' + serifs.reminder.notify(friend.name)
-				});
-			} catch (err) {
-				// renote対象が消されていたらリマインダー解除
-				if (err.statusCode === 400) {
-					this.unsubscribeReply(remind.thing == null && remind.quoteId ? remind.quoteId : remind.id);
-					this.reminds.remove(remind);
-					return;
-				}
+		} catch (err) {
+			// renote対象が消されていたらリマインダー解除
+			if (err.statusCode === 400) {
+				this.unsubscribeReply(remind.thing == null && remind.quoteId ? remind.quoteId : remind.id);
+				this.reminds.remove(remind);
 				return;
 			}
+			return;
 		}
 
-		this.subscribeReply(remind.id, remind.isDm, remind.isDm ? remind.userId : reply.id, {
+		this.subscribeReply(remind.id, reply.id, {
 			id: remind.id
 		});
 
