@@ -1,4 +1,4 @@
-import * as seedrandom from 'seedrandom';
+import seedrandom from 'seedrandom';
 
 export const itemPrefixes = [
 	'プラチナ製',
@@ -147,13 +147,11 @@ export const itemPrefixes = [
 	'先進的な',
 	'レトロな',
 	'ヴィンテージ',
-	'100日後に何らかが起きる',
 	'合法',
 	'プレミア付き',
 	'デカ',
 	'ギガ',
 	'穢れた',
-	'加護を受けた',
 	'品質保証付き',
 	'AppleCare+加入済み',
 	'えっちな',
@@ -200,6 +198,9 @@ export const itemPrefixes = [
 	'うねうね',
 	'水没',
 	'燃え盛る',
+	'高圧',
+	'異常',
+	'粗挽き',
 ];
 
 export const items = [
@@ -443,19 +444,19 @@ export const items = [
 	'オリハルコン',
 	'ヘドロ',
 	'グレーチング',
+	'繝九Λ縺ｮ縺ｿ縺晄ｱ',
+	'スーパーカミオカンデ',
+	'デースケドガー',
 ];
 
 export const and = [
 	'に擬態した',
 	'入りの',
-	'が埋め込まれた',
-	'を連想させる',
 	'っぽい',
 	'に見せかけて',
 	'を虐げる',
 	'を侍らせた',
 	'が上に乗った',
-	'のそばにある',
 ];
 
 export function genItem(seedOrRng?: (() => number) | string | number) {
